Refactor VirusTotal integration to use VTClient class and streamline asset processing
