+++ src/index.js
@@ -1,13 +1,8 @@
-const fs = require('node:fs')
-const path = require('node:path')
-
 const core = require('@actions/core')
 const github = require('@actions/github')
 const glob = require('@actions/glob')
 
-const { RateLimiter } = require('limiter')
-
-const vtUpload = require('./vt')
+const VTClient = require('./vt.js')
 
 ;(async () => {
     try {
@@ -17,25 +12,26 @@ const vtUpload = require('./vt')
         core.startGroup('Inputs')
         const inputs = getInputs()
         console.log(inputs)
-        core.endGroup() // Inputs
 
         // Set Variables
         const octokit = github.getOctokit(inputs.token)
         const release = await getRelease(octokit, inputs.release_id)
-        const limiter = new RateLimiter({
-            tokensPerInterval: inputs.rate,
-            interval: 'minute',
-        })
+        const client = new VTClient(inputs.key, inputs.rate)
+
+        core.endGroup() // Inputs
 
         // Process
         /** @type {Object[]} */
         let results
         if (inputs.files?.length) {
             // core.info('📁 Processing File Globs')
-            results = await processFiles(inputs, limiter)
+            const globber = await glob.create(inputs.files.join('\n'), {
+                matchDirectories: false,
+            })
+            results = await client.processFiles(await globber.glob())
         } else if (release) {
             // core.info('🗄️ Processing Release Assets')
-            results = await processRelease(inputs, limiter, octokit, release)
+            results = await client.processRelease(octokit, release.id)
         } else {
             return core.setFailed('No files or release to process.')
         }
@@ -135,7 +131,6 @@ const vtUpload = require('./vt')
 
         core.info('✅ \u001b[32;1mFinished Success')
     } catch (e) {
-        core.endGroup()
         console.log(e)
         if (e.response) {
             console.log(`\u001b[31m Error: ${e.response.statusText}`)
@@ -144,128 +139,6 @@ const vtUpload = require('./vt')
         core.setFailed(e.message)
     }
 })()
-
-/**
- * Process Release Assets
- * @param {Object} inputs
- * @param {RateLimiter} limiter
- * @param {InstanceType<typeof github.GitHub>} octokit
- * @param {Object} release
- * @return {Promise<Object[{id, name, link}]>}
- */
-async function processRelease(inputs, limiter, octokit, release) {
-    core.startGroup('Processing Release Assets')
-    core.startGroup('Release')
-    console.log(release)
-    core.endGroup() // Release
-
-    // Get Assets
-    let allAssets = []
-    let page = 0
-    const { data } = await octokit.rest.rateLimit.get()
-    const ghLimiter = new RateLimiter({
-        tokensPerInterval: data.resources.core.limit,
-        interval: 'hour',
-    })
-    while (true) {
-        await ghLimiter.removeTokens(1)
-        const assets = await octokit.rest.repos.listReleaseAssets({
-            ...github.context.repo,
-            release_id: release.id,
-            per_page: 100,
-            page: ++page,
-        })
-        if (!assets.data.length) break
-        allAssets = allAssets.concat(assets.data)
-    }
-    if (!allAssets.length) {
-        throw new Error(`No Assets Found for Release: ${release.id}`)
-    }
-
-    // Create Temp
-    console.log('RUNNER_TEMP:', process.env.RUNNER_TEMP)
-    const assetsPath = path.join(process.env.RUNNER_TEMP, 'assets')
-    console.log('assetsPath:', assetsPath)
-    if (!fs.existsSync(assetsPath)) {
-        fs.mkdirSync(assetsPath)
-    }
-
-    console.log(allAssets)
-    core.endGroup() // Assets
-
-    // Process Assets
-    const results = []
-    for (const asset of allAssets) {
-        core.startGroup(`Processing: \u001b[36m${asset.name}`)
-        if (inputs.rate) {
-            const remainingRequests = await limiter.removeTokens(1)
-            console.log('remainingRequests:', remainingRequests)
-        }
-        const filePath = path.join(assetsPath, asset.name)
-        console.log('filePath:', filePath)
-        const file = await octokit.rest.repos.getReleaseAsset({
-            ...github.context.repo,
-            asset_id: asset.id,
-            headers: {
-                Accept: 'application/octet-stream',
-            },
-        })
-        fs.writeFileSync(filePath, Buffer.from(file.data))
-        const result = await processVt(inputs, asset.name, filePath)
-        console.log('result:', result)
-        results.push(result)
-        core.endGroup() // Processing
-    }
-    return results
-}
-
-/**
- * Process File Globs
- * @param {Object} inputs
- * @param {RateLimiter} limiter
- * @return {Promise<Object[{id, name, link}]>}
- */
-async function processFiles(inputs, limiter) {
-    const globber = await glob.create(inputs.files.join('\n'), {
-        matchDirectories: false,
-    })
-    const files = await globber.glob()
-    core.startGroup('Processing File Globs')
-    console.log(files)
-    core.endGroup() // Files
-    if (!files.length) {
-        throw new Error('No files to process.')
-    }
-    const results = []
-    for (const file of files) {
-        const name = file.split('\\').pop().split('/').pop()
-        core.startGroup(`Processing: \u001b[36m${name}`)
-        if (inputs.rate) {
-            const remainingRequests = await limiter.removeTokens(1)
-            console.log('remainingRequests:', remainingRequests)
-        }
-        const result = await processVt(inputs, name, file)
-        // console.log('result:', result)
-        results.push(result)
-        core.endGroup() // Processing
-    }
-    return results
-}
-
-/**
- * Process VirusTotal
- * @param {Object} inputs
- * @param {String} name
- * @param {String} filePath
- * @return {Promise<{name, link: string, id}>}
- */
-async function processVt(inputs, name, filePath) {
-    const response = await vtUpload(filePath, inputs.key)
-    console.log('response.data.id:', response.data.id)
-    const link = `https://www.virustotal.com/gui/file-analysis/${response.data.id}`
-    console.log('link:', link)
-    return { id: response.data.id, name, link }
-}
 
 /**
  * Get Release
